* commit '3cdf705359afd3406429911a5f127c7d21a03599': (42 commits)
  PLGMAGTWOS-456: Bump version to 1.7.0
  PLGMAGTWOS-457: Update README.md
  PLGMAGTWOS-443: Bump version to 1.7.0-rc2
  PLGMAGTWOS-444: Update README
  PLGMAGTWOS-445: update CHANGELOG
  PLGMAGTWOS-431: Update payment methods spelling
  PLGMAGTWOS-288: Add payment specification when paid with coupon and payment method
  PLGMAGTWOS-435: [INTERNAL] improve ParseAddress function
  PLGMAGTWOS-429: [INTERNAL] Update translation for PLGMAGTWOS-350
  PLGMAGTWOS-302: Changed VVV Bon to VVV Cadeaukaart
  PLGMAGTWOS-418: Disabled second chance mail on backend orders
  PLGMAGTWOS-371: Extract setMultisafepayStatus so it can be used by plugins. Thanks to [Tjitse-E](https://github.com/Tjitse-E)
  PLGMAGTWOS-408: Correct spelling of canceled in MultiSafepay Settings
  PLGMAGTWOS-383: Remove support for days and seconds active for bank transfer
  PLGMAGTWOS-301: Disable manually created payment links for Klarna, Pay After Delivery, AfterPay, Santander Betaalplan
  PLGMAGTWOS-364: Changed license.md content from php comment to markdown
  PLGMAGTWOS-409: Added DISCLAIMER.md
  PLGMAGTWOS-423: Bank transfer does not pass validation
  PLGMAGTWOS-422: Changed the author of the disclaimer
  PLGMAGTWOS-404: Add payment link to order confirmation email
  ...

+++ Controller/Connect/Notification.php
@@ -17,7 +17,7 @@
  *
  * @category    MultiSafepay
  * @package     Connect
- * @author      Ruud Jonk <techsupport@multisafepay.com>
+ * @author      MultiSafepay <techsupport@multisafepay.com>
  * @copyright   Copyright (c) 2018 MultiSafepay, Inc. (https://www.multisafepay.com)
  * @license     http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  *
@@ -104,6 +104,10 @@ class Notification extends \Magento\Framework\App\Action\Action
     {
         $params = $this->_requestHttp->getParams();
 
+        if(!$this->validateParams($params)){
+            return false;
+        }
+
         if(isset($params['hash'])){
             $recurringId = $this->_mspHelper->getRecurringIdByHash($params['hash']);
             $this->_mspToken->create()->load($recurringId)->delete();
@@ -126,7 +130,6 @@ class Notification extends \Magento\Framework\App\Action\Action
                 $paymentMethod->_invoiceSender = $this->_invoiceSender;
                 $storeManager = $this->_storeManager;
                 $paymentMethod->_stockInterface = $this->_stockRegistryProvider;
-
 
                 $updated = $paymentMethod->notification($order);
                 $this->_mspHelper->unlockProcess(
@@ -158,4 +161,15 @@ class Notification extends \Magento\Framework\App\Action\Action
             $this->getResponse()->setContent('Order not found');
         }
     }
+
+    /**
+     * @param array $params
+     *
+     * @return bool
+     */
+    private function validateParams($params)
+    {
+        return isset($params['transactionid']) && is_numeric($params['transactionid']);
+    }
+
 }
